requirements.txt
Flask
youtube-transcript-api
openai 
python-dotenv 